Server tests for scenarios
