+++ server/BimsAPI/BimsAPI.test.ts
@@ -3,7 +3,6 @@ import supertest, {Response} from "supertest";
 
 import MockAdapter from "axios-mock-adapter";
 import axios from "axios";
-import {defineFeature, loadFeature} from "jest-cucumber";
 
 // Mock Express Server
 const request = supertest(app);
@@ -57,58 +56,5 @@ describe("Data Delivery Get all batches from API", () => {
         jest.clearAllMocks();
         jest.resetModules();
         mock.reset();
-    });
-});
-
-const feature = loadFeature("./src/features/set_telephone_operations_start_date.feature", {tagFilter: "@server"});
-
-/**
- * These scenarios don't really work as unit tests,
- * but they work as a decent structure to unit test the BimsAPI requests in the node.js server.
- */
-defineFeature(feature, test => {
-
-    afterEach(() => {
-        jest.clearAllMocks();
-        jest.resetModules();
-        mock.reset();
-    });
-
-    test("Overwrite questionnaire and previous TO Start Date with new", ({given, when, then}) => {
-        let response: Response;
-
-        given("a pre-deployed questionnaire that already has a TO Start Date stored against it", () => {
-            // Mock BIMS already having a live date stored
-            mock.onGet(/\/tostartdate\/OPN2004A$/).reply(200, {"tostartdate": "2021-07-29T00:00:00+00:00"}, jsonHeaders);
-            // Mock BIMS storing updated date
-            mock.onPatch(/\/tostartdate\/OPN2004A$/).reply(200, [], jsonHeaders);
-        });
-
-        when("a TO Start Date is specified", async () => {
-            response = await request.post("/api/tostartdate/OPN2004A").send({"tostartdate": "2020-06-05"});
-        });
-
-        then("the new TO Start Date will overwrite the previous", () => {
-            expect(response.status).toEqual(200);
-        });
-    });
-
-    test("Overwrite questionnaire and remove previous TO Start Date", ({given, when, then}) => {
-        let response: Response;
-
-        given("a pre-deployed questionnaire that already has a TO Start Date stored against it", () => {
-            // Mock BIMS already having a live date stored
-            mock.onGet(/\/tostartdate\/OPN2004A$/).reply(200, {"tostartdate": "2021-07-29T00:00:00+00:00"}, jsonHeaders);
-            // Mock BIMS deleting date
-            mock.onDelete(/\/tostartdate\/OPN2004A$/).reply(204, [], jsonHeaders);
-        });
-
-        when("a TO Start Date is not specified", async () => {
-            response = await request.post("/api/tostartdate/OPN2004A").send({"tostartdate": ""});
-        });
-
-        then("the original TO Start Date is removed from data store", () => {
-            expect(response.status).toEqual(200);
-        });
     });
 });
