+++ src/features/set_telephone_operations_start_date.feature
@@ -33,8 +33,7 @@ Feature: DQS - Set a TO Start Date during deployment
     Then I am presented with an information banner with an error message
 
   #  Scenario 5:
-  # This logic is handled in the Node.js Sever so tests are there in :
-  # TODO: ADD Filename here
+  # This logic is handled in the Node.js Sever so tests are there in: ./server/BimsAPI/set_to_start_date_server.test.ts
   @server
   Scenario: Overwrite questionnaire and previous TO Start Date with new
     Given a pre-deployed questionnaire that already has a TO Start Date stored against it
@@ -42,8 +41,7 @@ Feature: DQS - Set a TO Start Date during deployment
     Then the new TO Start Date will overwrite the previous
 
   #  Scenario 6:
-  # This logic is handled in the Node.js Sever so tests are there in :
-  # TODO: ADD Filename here
+  # This logic is handled in the Node.js Sever so tests are there in: ./server/BimsAPI/set_to_start_date_server.test.ts
   @server
   Scenario: Overwrite questionnaire and remove previous TO Start Date
     Given a pre-deployed questionnaire that already has a TO Start Date stored against it
